osx/requirements.txt
numpy
scikit-image
scipy
scikit-learn
smplx==0.1.28
tqdm
yacs
numba
opencv-python
tensorboardx
filterpy
cython
chumpy
Pillow
trimesh
pyrender
matplotlib
json_tricks
torchgeometry
einops
joblib
boto3
requests
easydict
pycocotools
plyfile
timm
pyglet==1.5.27
